A number of credentials improvements
