+++ CHANGELOG.md
@@ -16,6 +16,12 @@ Changes since the last release
 -   Added CredentialGenerator base generator class to implement cached credential generators
 -   Added GeneratorContext class to implement context validation
 -   Updated some builtin generators to avoid redefining \_\_init\_\_ making them more flexible
+-   Updated the CredentialPair class and its subclasses to support credential pairs with different types. (PR #514)
+-   Refactored the credentials module into a python package including modules for base classes and specific credential types. (PR #514)
+-   Renamed the credentialGenerator and hashCrypto modules to credential_generator and hash_crypto. (PR #514)
+-   Renamed CredentialGenerator credentials to DynamicCredential and added the new alias "generator" to the newly renamed "dynamic" credential type. (PR #514)
+-   Added the new RSAPublicKey, RSAPrivateKey, and Symmetric credential types to replace the legacy pubCrypto and symCrypto modules. (PR #514)
+-   Stopped using M2Crypto in favor of the cryptography module. (PR #514)
 
 ### Changed defaults / behaviours
 
@@ -24,6 +30,8 @@ Changes since the last release
 ### Security Related Fixes
 
 ### Bug Fixes
+
+-   Fixed a bug that could cause the Frontend not to generate IDTOKENS to some Factory entries. (PR #514)
 
 ### Testing / Development
