+++ factory/glideFactoryConfig.py
@@ -135,7 +135,7 @@ class JoinConfigFile(ConfigFile):
 ############################################################
 
 
-class GlideinKey:  # TODO: Check for credentials refactor
+class GlideinKey:  # NOTE: This class doesn't seem to be used anywhere
     def __init__(self, pub_key_type, key_fname=None, recreate=False):
         self.pub_key_type = pub_key_type
         self.load(key_fname, recreate)
@@ -247,7 +247,7 @@ class GlideinDescript(ConfigFile):
 
         if self.data["OldPubKeyType"] is not None:
             try:
-                self.data["OldPubKeyObj"] = credentials.RSAKey(path=self.backup_rsakey_fname)
+                self.data["OldPubKeyObj"] = credentials.RSAPrivateKey(path=self.backup_rsakey_fname)
             except Exception:
                 self.data["OldPubKeyType"] = None
                 self.data["OldPubKeyObj"] = None
@@ -271,7 +271,7 @@ class GlideinDescript(ConfigFile):
             recreate (bool): Create a new key overwriting the old one. Defaults to False
         """
         if self.data["PubKeyType"] is not None:
-            self.data["PubKeyObj"] = credentials.RSAKey(path=self.default_rsakey_fname)
+            self.data["PubKeyObj"] = credentials.RSAPrivateKey(path=self.default_rsakey_fname)
             if recreate:
                 self.data["PubKeyObj"].recreate()
         else:
