+++ factory/glideFactoryInterface.py
@@ -625,7 +625,7 @@ class EntryClassad(classadSupport.Classad):
         if pub_key_obj is not None:
             self.adParams["PubKeyID"] = "%s" % pub_key_obj.pub_key_id
             self.adParams["PubKeyType"] = "%s" % pub_key_obj.key_type
-            self.adParams["PubKeyValue"] = "%s" % pub_key_obj.pub_key.get().decode("ascii").replace("\n", "\\n")
+            self.adParams["PubKeyValue"] = "%s" % pub_key_obj.pub_key.string.decode("ascii").replace("\n", "\\n")
         if "grid_proxy" in auth_method:  # TODO: Check for credentials refactoring impact
             self.adParams["GlideinAllowx509_Proxy"] = "%s" % True
             self.adParams["GlideinRequirex509_Proxy"] = "%s" % True
@@ -709,7 +709,7 @@ class FactoryGlobalClassad(classadSupport.Classad):
         self.adParams["GlideinWMSVersion"] = factoryConfig.glideinwms_version
         self.adParams["PubKeyID"] = "%s" % pub_key_obj.pub_key_id
         self.adParams["PubKeyType"] = "%s" % pub_key_obj.key_type
-        self.adParams["PubKeyValue"] = "%s" % pub_key_obj.pub_key.get().decode("ascii").replace("\n", "\\n")
+        self.adParams["PubKeyValue"] = "%s" % pub_key_obj.pub_key.string.decode("ascii").replace("\n", "\\n")
 
 
 def advertiseGlobal(
