+++ lib/generators/credential_generator.py
